Refactor: Decompose `relvar-core/src/database/mod.rs` into submodules
- Extracted `VirtualRelvarDefinition` to `relvar-core/src/database/virtual_relvar.rs`.
- Extracted `compute_relation_after_delete` and `compute_relation_after_update` to `relvar-core/src/database/dml.rs`.
- Updated `relvar-core/src/database/mod.rs` to use these new modules.
- Re-exported `VirtualRelvarDefinition` in `mod.rs` to maintain API compatibility.

This improves cohesion and reduces the size of `database/mod.rs`.
